# cmd/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
...
    "/admin/payment/payment-method/new": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "admin can add new payment methods",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Admin"],
        "summary": "ADD NEW PAYMENT METHOD",
        "parameters": [
          {
            "description": "payment method",
            "name": "payment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.NewPaymentMethod"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "models.NewPaymentMethod": {
      "type": "object",
      "properties": {
        "payment_method": {
          "type": "string"
        }
      }
    },
...
    "response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "error": {},
        "message": {
          "type": "string"
        },
        "status_code": {
          "type": "integer"
        }
      }
    }
  }
}